fix: Fixed map to placing bug
Fixed a bug where switching to a project with less points would cause an error, because the selectedPoint for placing would be out of range in the project. The solution sets the placed point to -1 when switching projects, shows a special message that forces the user to select a point before placing and uses the points in projectSettings directly, rather then using them from placingSettings.